Add some statistical methods to RRi class

+++ hrv/rri.py
@@ -39,7 +39,7 @@ class RRi:
 
 
 def _validate_rri(rri):
-    rri = np.array(rri)
+    rri = np.array(rri, dtype=np.float32)
 
     if any(rri <= 0):
         raise ValueError('rri series can only have positive values')
@@ -52,7 +52,7 @@ def _validate_rri(rri):
 
 
 def _validate_time(rri, time):
-    time = np.array(time)
+    time = np.array(time, dtype=np.float32)
     if len(rri) != len(time):
         raise ValueError('rri and time series must have the same length')
